+++ src/arc/_listener.js
@@ -1,0 +1,23 @@
+let _ssm = require('./_ssm')
+
+module.exports = function _arcListener ({ inventory }, req, res) {
+  if (req.method.toLowerCase() === 'post') {
+    let body = ''
+
+    req.on('data', chunk => {
+      body += chunk.toString()
+    })
+
+    req.on('end', () => {
+      if (req.url === '/_arc/ssm') {
+        _ssm({ inventory, body }, req, res)
+        return
+      }
+    })
+  }
+  else {
+    res.statusCode = 404
+    res.end()
+    return
+  }
+}